backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain-openai==0.0.5
python-dotenv==1.0.0
pydantic==1.10.12
pymongo==4.5.0
email-validator==2.1.0